refactor: 불필요한 계층 제거
