--- src/main/java/janggi/piece/UnjumpablePiece.java
@@ -1,57 +1,0 @@
-package janggi.piece;
-
-import janggi.Board;
-import janggi.MoveVector;
-import janggi.MovingRule;
-import janggi.Position;
-import java.util.List;
-
-public abstract class UnjumpablePiece extends Piece {
-
-    public UnjumpablePiece(final Team team, final List<MovingRule> movingRules) {
-        super(team, movingRules);
-    }
-
-    @Override
-    public boolean canJump() {
-        return false;
-    }
-
-    @Override
-    public boolean canMove(final Position start, final Position end, final Board board) {
-        if (cannotFindRule(start, end)) {
-            return false;
-        }
-        if (cannotMoveThrough(start, end, board)) {
-            return false;
-        }
-        return !isPresentSameTeam(end, board);
-    }
-
-    private boolean isPresentSameTeam(final Position end, final Board board) {
-        return board.isPresentSameTeam(team, end);
-    }
-
-    private boolean cannotMoveThrough(final Position start, final Position end, final Board board) {
-        final MovingRule matchRule = findMatchRule(start, end);
-        Position route = start;
-        for (MoveVector vector : matchRule.getVectorsWithoutLast()) {
-            route = route.add(vector);
-            if (board.isPresent(route)) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    private MovingRule findMatchRule(final Position start, final Position end) {
-        final MoveVector startEndDiff = end.getDiff(start);
-        for (MovingRule movingRule : movingRules) {
-            final MoveVector vectorSum = movingRule.sumUnit();
-            if (vectorSum.equals(startEndDiff)) {
-                return movingRule;
-            }
-        }
-        throw new IllegalStateException();
-    }
-}
